Reorganize worktop
