+++ radix-engine/src/model/substates/mod.rs
@@ -6,6 +6,7 @@ mod resource_manager;
 mod substate;
 mod system;
 mod vault;
+mod worktop;
 
 pub use component::*;
 pub use kv_store_entry::*;
@@ -15,3 +16,4 @@ pub use resource_manager::*;
 pub use substate::*;
 pub use system::*;
 pub use vault::*;
+pub use worktop::*;
